head/requirements.txt
flask>=2.0.0
hydra-core>=1.1.0
omegaconf>=2.1.0
opencv-python>=4.5.0
adafruit-circuitpython-servokit>=1.3.0
numpy>=1.20.0